remove some unused files and app new files to pool
