--- poc/errors.py
@@ -1,7 +1,0 @@
-class PoolClosed(Exception): ...
-
-
-class PoolTimeout(Exception): ...
-
-
-class PoolAlreadyStarted(Exception): ...
