+++ pysqlx_engine/_core/conn.pyi
@@ -1,8 +1,8 @@
 from types import TracebackType
 from typing import Any, Dict, List, Optional, Type, Union, overload
 
-from .._core.parser import BaseRow, MyModel, DictParam  # import necessary using _core to not subscribe default parser
 from .const import ISOLATION_LEVEL
+from .parser import BaseRow, DictParam, MyModel
 
 class PySQLXEngineSync:
 	"""
@@ -950,4 +950,6 @@ class PySQLXEngineSync:
 		    * [MySQL](https://dev.mysql.com/doc/refman/8.0/en/innodb-transaction-isolation-levels.html)
 		    * [SQLite](https://www.sqlite.org/isolation.html)
 		"""
+		...
+		...
 		...
