+++ pysqlx_engine/_core/util.py
@@ -1,11 +1,13 @@
 import asyncio
 import functools
 import shutil
-from datetime import date, datetime, time
+import threading
+import time
+from datetime import date, datetime
 from decimal import Decimal
 from enum import Enum
 from functools import lru_cache
-from typing import Callable, List, TypeVar, Union
+from typing import Any, Callable, Coroutine, List, TypeVar, Union
 from uuid import UUID
 
 from pydantic import BaseModel
@@ -138,3 +140,45 @@ def create_log_line(text: str, character: str = "=") -> str:
 		line += character * (terminal_width - len(line))  # pragma: no cover
 
 	return line
+
+
+T = TypeVar("T")
+
+
+def aspawn(
+	f: Callable[..., Coroutine[Any, Any, None]], args: tuple[Any, T] = (), name: str | None = None
+) -> asyncio.Task[None]:
+	"""
+	Equivalent to asyncio.create_task.
+	"""
+	return asyncio.create_task(f(*args), name=name)
+
+
+def asleep(seconds: float) -> Coroutine[Any, Any, None]:
+	"""
+	Equivalent to asyncio.sleep(), converted to time.sleep() by async_to_sync.
+	"""
+	return asyncio.sleep(seconds)
+
+
+def sleep(seconds: float) -> None:
+	"""
+	Equivalent to time.sleep().
+	"""
+	return time.sleep(seconds)
+
+
+def spawn(f: Callable[..., Any], args: tuple[Any, T] = (), name: str | None = None) -> threading.Thread:
+	"""
+	Equivalent to creating and running a daemon thread.
+	"""
+	t = threading.Thread(target=f, args=args, name=name, daemon=True)
+	t.start()
+	return t
+
+
+def get_task_name(task: Union[asyncio.Task, threading.Thread]) -> str:
+	if isinstance(task, asyncio.Task):
+		return task.get_name()
+
+	return task.name
